MP-04: retire copilot-backend.ts, the direct Arch A builder
Arch B (BYOCA/MCP) and Arch C (managed agents on Anthropic/Gemini/Copilot)
are the production paths now; the direct Copilot backend that dispatches
through the Agent Tasks API and stages/deletes git branches has no live
traffic left to serve. Per the ops-repo migration plan (MP-01...MP-04),
retiring it means:

- Delete copilot-backend.ts. Its seed-staging and cleanup logic already
  moved into the shared agent-tasks.ts helper (this branch's earlier
  commit), so managed-provider-copilot.ts loses nothing by this deletion
  — it was already the sole caller of that shared logic going forward.
- agent-backend-env.ts: remove createPlatformBackendFromEnv and the
  deprecated createAgentBackendFromEnv wrapper (zero remaining callers).
  createAgentBackendRegistryFromEnv no longer falls back to the legacy
  backend when MANAGED_AGENT_VENDOR is unset or invalid — an unconfigured
  environment now correctly has no platform builder rather than a silent
  Copilot dispatch nobody asked for.
- agent-backend-env.test.ts: updated the one test that asserted the old
  fallback behavior to assert its absence instead.

copilot-backend.test.ts was 29 tests, not all of them about the backend:
- The 11 `buildPrompt` tests and the 1 buildPrompt-only seeded-dispatch
  test exercised build-prompt.ts's default (harness-lane) contract, not
  the backend. Moved verbatim into build-prompt.test.ts.
- The 2 `dispatch through the submissions route` tests exercised
  store.ts's recordDispatch/getSubmission, generic to any backend name.
  Moved verbatim into store.test.ts.
- The remaining 16 tests (dispatch/resume/cleanup/observe+cancel/seeded
  dispatch against the backend itself) are deleted along with the
  backend, on the basis that managed-provider-copilot.test.ts already
  covers the equivalent behavior for the surviving Copilot driver:
  pinned model + no PR, branch-from-task reporting, credit summation,
  seed staging on the same branch naming, fail-open on seed failure,
  cancellation honesty, and workspace deletion. Two things do NOT have
  an exact equivalent and are called out rather than silently assumed
  ported: copilot-backend.ts's resume() had its own doc comment on why
  a revision starts a fresh branch rather than continuing the stale one
  (managed-provider-copilot.ts has no resume() of its own — that
  semantics lives in managed-backend.ts, untouched here), and the
  "does nothing for a dispatch with no workspace" guard on cleanup()
  doesn't apply the same way to deleteWorkspace()'s required-string
  signature.

Also fixes one stale reference in docs/build-brief.md that named
copilot-backend.test.ts as one of two suites pinning the untrusted-spec
contract; that job now belongs to build-prompt.test.ts alone.

Full gate green: lint (incl. comment-prose), type-check, the complete
test suite (12+96+3112+1288+20+36 = 4564 tests across the monorepo,
API workspace at 3112 after relocating/removing 15 net tests), build.

Not done here, and still gated per the ops-repo plan: the availability
switch (managed-availability.ts) still selects only on/off, not a driver
dimension — MP-04's brief calls for driver-level select/failover from
the ops-config document without a redeploy. Out of scope for this PR;
flagging so it isn't mistaken for already covered.
