+++ apps/api/src/agent-backend-env.ts
@@ -6,10 +6,7 @@
 
 import type { AgentBackend } from './agent-backend.js';
 import type { BuilderKind } from './builder.js';
-import { createAgentTasksClient, type AgentTaskModel } from './agent-tasks.js';
-import { createCopilotBackend } from './copilot-backend.js';
 import { VertexGameSeeder, type GameSeeder } from './game-seed.js';
-import { createGitHubClient } from './github-client.js';
 import { createManagedProvider, type ManagedAgentEffort, type ManagedPromptLane } from './managed-agent.js';
 import './managed-provider-anthropic.js';
 import './managed-provider-copilot.js';
@@ -40,15 +37,10 @@ export function resolveBuilderBackend(registry: AgentBackendRegistry, builder: B
   return builder === 'self' ? registry.self : registry.platform;
 }
 
-/**
- * Returns a platform backend, or undefined when this environment is not set up to
- * dispatch Copilot.
- *
- * Undefined is a supported state, not a failure: local development has no dispatch
- * credential, and an environment can run the whole product without one — self builds
- * still work, and submissions that ask for the platform builder wait in `queued`.
- */
 // What the managed backend needs that the environment cannot supply.
+//
+// A missing platform backend (undefined) is a supported state, not a failure: local
+// development has no dispatch credential, and submissions asking for it wait in `queued`.
 export interface ManagedBackendDeps {
   deliver?: ManagedDeliverySink;
   lock?: ManagedDeliveryLock;
@@ -211,38 +203,11 @@ export function createManagedPlatformBackendFromEnv(deps?: ManagedBackendDeps, l
   });
 }
 
-export function createPlatformBackendFromEnv(log?: Logger): AgentBackend | undefined {
-  // Deliberately its own credential rather than reusing GITHUB_TOKEN. The agent tasks
-  // API needs a user-to-server token (installation tokens are unsupported), and keeping
-  // it separate means dispatch and serving fail independently — a dispatch PAT expiring
-  // must not take the catalog down with it, which is exactly the shared-budget coupling
-  // that 403'd two jobs at once on 2026-07-28.
-  const token = process.env.AGENT_TASKS_TOKEN?.trim();
-  const repo = process.env.GAMES_REPO?.trim() ?? 'gamedevpl/www.gamedev.pl-games';
-  if (!token) return undefined;
-
-  const model = (process.env.AGENT_TASKS_MODEL?.trim() as AgentTaskModel | undefined) ?? 'claude-sonnet-4.6';
-
-  log?.info({ repo, model, backend: 'copilot' }, 'agent dispatch enabled');
-
-  return createCopilotBackend({
-    tasks: createAgentTasksClient({ token, repo }),
-    // Deletes spent workspaces, and commits a generated seed onto the branch a seeded
-    // build starts from.
-    github: createGitHubClient({ token, repo }),
-    baseRef: process.env.GAMES_PUBLISHED_REF?.trim() || 'main',
-    model,
-    customAgent: process.env.AGENT_CUSTOM_AGENT?.trim() || 'game-builder',
-    // Delivery is by upload; a pull request is opened lazily, only when a revision round
-    // needs one as resumption context.
-    createPullRequest: false,
-    ...(log ? { log } : {}),
-  });
-}
-
 /**
- * Builds the registry. Self is always available (no external credential); platform
- * follows {@link createPlatformBackendFromEnv}.
+ * Builds the registry. Self is always available (no external credential); platform is
+ * whatever {@link createManagedPlatformBackendFromEnv} resolves — the direct Copilot
+ * backend (`copilot-backend.ts`) was retired in MP-04, so an unset or invalid
+ * `MANAGED_AGENT_VENDOR` now means no platform builder, not a silent Copilot fallback.
  */
 export function createAgentBackendRegistryFromEnv(
   log?: Logger,
@@ -250,24 +215,14 @@ export function createAgentBackendRegistryFromEnv(
   managedDeps?: ManagedBackendDeps,
 ): AgentBackendRegistry {
   const selectedVendor = process.env.MANAGED_AGENT_VENDOR?.trim();
-  const managed = createManagedPlatformBackendFromEnv(managedDeps, log);
-  // Explicit vendor must not silently fall back to Copilot.
-  const platform = selectedVendor ? managed : (managed ?? createPlatformBackendFromEnv(log));
+  const platform = createManagedPlatformBackendFromEnv(managedDeps, log);
   const self = createSelfBuildBackend(selfOptions);
-  if (selectedVendor && !managed) {
+  if (selectedVendor && !platform) {
     log?.warn({ vendor: selectedVendor }, 'managed agent vendor is set but invalid; platform dispatch stays off');
   } else if (!platform) {
     log?.info({ backend: 'self' }, 'self-build backend enabled (no platform dispatch credential)');
   }
   return { ...(platform ? { platform } : {}), self };
-}
-
-/**
- * @deprecated Prefer {@link createAgentBackendRegistryFromEnv}. Kept so older call sites
- * that expect a single backend still resolve the platform adapter.
- */
-export function createAgentBackendFromEnv(log?: Logger): AgentBackend | undefined {
-  return createPlatformBackendFromEnv(log);
 }
 
 /**
