+++ apps/api/src/store.test.ts
@@ -655,3 +655,30 @@ describe('creator message history', () => {
     ]);
   });
 });
+
+// Relocated from copilot-backend.test.ts (MP-04) — generic store behavior.
+describe('dispatch records', () => {
+  it('is skipped without a backend, so a submission never fails on missing orchestration', async () => {
+    // A missing dispatch credential must not lose the submission itself.
+    const store = new InMemoryStore();
+    await store.createSubmission(1, 'g:1', 'A game');
+
+    expect((await store.getSubmission(1))?.dispatch).toBeUndefined();
+  });
+
+  it('records every round against one workspace', async () => {
+    // Refs accumulate per round; the workspace and newest ref stay current.
+    const store = new InMemoryStore();
+    await store.createSubmission(1, 'g:1', 'A game');
+
+    await store.recordDispatch(1, { backend: 'copilot', ref: 'task-1', workspace: 'copilot/x' });
+    await store.recordDispatch(1, { backend: 'copilot', ref: 'task-2' });
+
+    expect((await store.getSubmission(1))?.dispatch).toEqual({
+      backend: 'copilot',
+      refs: ['task-1', 'task-2'],
+      workspace: 'copilot/x',
+      seedWorkspace: undefined,
+    });
+  });
+});
